# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 0.0.1
  title: complex API
host: "localhost:20016"
basePath: /Adam/complex
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
  - text/plain
paths:
...
  /count:
    x-swagger-router-controller: controller
    get:
      description: returns count of 'complex'
      operationId: v1_srvc2006Count
      parameters:
        - &ref_0
          name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - &ref_1
          name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - &ref_2
          name: expand
          in: query
          type: boolean
          description: expand document based on relations
          default: false
      responses:
        "200":
          description: Count of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
...
  /utils/fileTransfers:
    x-swagger-router-controller: controller
    get:
      description: Retrieve list of bulk actions'
      operationId: v1_srvc2006ExportDetails
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  /utils/fileTransfers/count:
    x-swagger-router-controller: controller
    get:
      description: count of bulk actions'
      operationId: v1_srvc2006ExportDetailsCount
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  "/utils/fileTransfers/{id}":
    x-swagger-router-controller: controller
    delete:
      description: delete file'
      operationId: v1_srvc2006ExportDetailsDelete
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Id of file
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  /utils/aggregate:
    x-swagger-router-controller: controller
    post:
      description: runs aggregate query
      operationId: v1_aggregate
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - name: data
          in: body
          description: Payload to aggregate
          schema:
            type: array
            items:
              type: object
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  /utils/hrefUpdate:
    x-swagger-router-controller: controller
    put:
      description: Update url of outgoing relations
      operationId: v1_updateHref
      parameters:
        - name: data
          in: body
          description: Payload to update href
          schema:
            type: object
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
...